Moved random video logic from videogen to narreddit
